0.0.2

+++ app/src/main/java/com/example/weatherapp/data/api/forecastEntities/CityWeatherForecast.kt
@@ -2,10 +2,10 @@ package com.example.weatherapp.data.api.forecastEntities
 
 data class CityWeatherForecast(
     val clouds: Clouds,
-    val dt: Int,
+    val dt: Long,
     val dt_txt: String,
     val main: Main,
-    val pop: Int,
+    val pop: Float,
     val sys: Sys,
     val visibility: Int,
     val weather: List<Weather>,
